aer-inject: support 32-bit domains
Domains can be up to 32-bits in length, and are supported in the linux-
kernel as of commit 28ef241f.

+++ aer.h
@@ -16,7 +16,7 @@ struct aer_error_inj
 	int32_t header_log1;
 	int32_t header_log2;
 	int32_t header_log3;
-	uint16_t domain;
+	uint32_t domain;
 };
 
 #define  PCI_ERR_UNC_TRAIN	0x00000001	/* Training */
